Terminamos mp y acomodamos casi todo

+++ src/main/java/com/tallerwebi/infraestructura/servicio/ServicioMercadoPago.java
@@ -2,5 +2,5 @@ package com.tallerwebi.infraestructura.servicio;
 
 public interface ServicioMercadoPago {
 
-    String realizarPago(Long cantidad);
+    String realizarPago(Integer cantidad);
 }
